Correct problem with prune users when smf used as forum: Error: You have an error in your SQL syntax; check the manual that corresponds to your MySQL server version for the right syntax to use near ')' at line 1 Query: DELETE FROM smf_members WHERE ID_MEMBER IN()

+++ admin/admin.prune_users.php
@@ -59,12 +59,12 @@ if ($action=="prune")
      if($GLOBALS["FORUMLINK"]=="smf")
      {
          $smf_fid=array();
-         foreach($_POST["smf_fid"] AS $v)
+         foreach($_POST["smf_fid"] AS $id=>$v)
          {
              $smf_fid[]=intval($v);
          }
-        
-         do_sqlquery("DELETE FROM {$db_prefix}members WHERE ID_MEMBER IN(".implode(",", $smf_fid).")",true);
+         if (implode("", $smf_fid)!="")
+            do_sqlquery("DELETE FROM {$db_prefix}members WHERE ID_MEMBER IN (".implode(",", $smf_fid).")",true);
      }
      
      $block_title=$language["PRUNE_USERS_PRUNED"];
@@ -102,6 +102,8 @@ elseif ($action=="view")
              $ru[$i]["id"]=$rusers["id"];
              if($GLOBALS["FORUMLINK"]=="smf")
                  $ru[$i]["smf_fid"]=$rusers["smf_fid"];
+             else
+                 $ru[$i]["smf_fid"]="";
              $i++;
            }
       unset($res);
